refactor: split client state from notes viewer
