+++ skills/html-prototype-build/addons/annotations/ADDON.md
@@ -8,14 +8,15 @@
 
 - 稳定目标锚点，优先使用元素现有的 `id`。
 - `window.__PROTOTYPE_NOTES__` snapshot。
-- `runtime/client/notes/viewer.js`。
+- `runtime/client/core/state.js` 提供 `PrototypeViewers`。
+- `runtime/client/notes/viewer.js` 提供只读说明渲染；正式交付同时加载 `runtime/client/core/display-mode.js` 处理 `product-only`。
 - snapshot 使用显式 `scenarios`，深链使用 `?scene=<id>`。
 
 ## 边界
 
 - 本 Addon 不提供产品导航、表单、数据或反馈组件。
 - 标注层专用视觉不得写入所选 UI 包的产品 Token。
-- Author Tools 的 Mark、Notes Editor、Inspector 和本地服务不属于本 Addon。
+- Mark 评审 pin、Notes Editor、Inspector 和本地服务不属于本 Addon。
 - Addon 只消费当前状态决定说明展示，不得新增业务状态源。
 - 不为状态新增 `data-*`；无合适 `id` 时才使用 `data-prototype-note-target` 作为锚点兜底。
