+++ skills/html-prototype-build/addons/annotations/ui-annotations.html
@@ -1,9 +1,9 @@
 <!--
-  UI 正式标注包：所有原型均加载，作为 Viewer 与统一状态协调器的运行时入口。
-  右栏 DOM、样式和连线逻辑由本 Skill 的 runtime/viewer.js 运行时创建，
+  UI 正式标注包：所有原型均加载，作为 Notes Viewer 与统一状态协调器的声明式示例。
+  状态协调由 runtime/client/core/state.js 提供，右栏 DOM、样式和连线逻辑由 runtime/client/notes/viewer.js 创建；
   原型 HTML 不得硬编码右栏结构、卡片正文或连线逻辑。
   卡片只展示标题与正文；target.label 保留为结构化数据，不在卡片内重复显示。
-  作者态操作由 html-prototype-build 本地服务（serve.mjs）注入到卡片右上角和面板底部。
+  作者态操作由 html-prototype-build 本地服务（runtime/server/index.mjs）动态注入。
 -->
 
 <!-- 原型优先复用语义 id 作为稳定锚点；状态不存入 DOM 属性。 -->
@@ -65,6 +65,8 @@ window.__PROTOTYPE_NOTES__ = {
 };
 </script>
 
-<!-- prototype.html 末尾只保留以下只读引用；Viewer 同时提供 PrototypeViewers 状态协调器。 -->
+<!-- prototype.html 末尾按 Client Runtime 固定顺序加载。 -->
 <script src="./prototype/notes.snapshot.js"></script>
+<script src="./prototype/display-mode.js"></script>
+<script src="./prototype/state.js"></script>
 <script src="./prototype/viewer.js" data-prototype-notes-viewer></script>
